move devnet to seperate file
